Plotting 'networktree' objects

Wraps plot.party to plot a tree model with networks on the ends. Networks are plotted with qgraph, and additional arguments are passed there

# S3 method for networktree
plot(
  x,
  terminal_panel = NULL,
  transform = NULL,
  layout = "lock",
  sdbars = FALSE,
  tnex = 3,
  partyargs = list(),
  na.rm = TRUE,
  ...
)

Arguments

x

an object of type 'networktree'

terminal_panel

an optional panel function of the form function(node) plotting the terminal nodes. Alternatively, a panel generating function of class "grapcon_generator" that is called with arguments x and tp_args to set up a panel function. Or, a character choosing one of the implemented standard plots "graph", "box", "matrix" or "bar". The default (NULL) chooses an appropriate panel function depending on the "model" argument.

transform

"cor", "pcor", or "glasso". If set to NULL, transform detected from x

layout

network layout, passed to qgraph. Default "lock" computes spring layout for the full sample and applies this to all graphs

sdbars

if using a barplot, should std deviation error bars be plotted?

tnex

terminal node extension (passed to plot.party). To make the terminal plots bigger, increase this value.

partyargs

additional arguments (list format) passed to partykit::plot.party plotting function that takes partitioned data as input

na.rm

should NA values be removed prior to calculating relevant parameters?

...

additional arguments passed to qgraph or barplot